Added the expiration date to the Add job/meetup form
I've added the expiration date to the Add job/meetup form so the person posting the offer can
specify till when the post should be published. I've made this field not required.

I've also changed the datetime widget used in the Add meetup for to bootstrap 3 datetime widget.
The reason for that was that the previous datetime widget had no option to display just a calendar
without collecting the time. And as the new widget doesn't go along well with the previous one,
I've also changed the datetime fields to use the new widget.

@ksiedlarek, I hope it's OK with you.